fix: checking for extensions as well before reusing the vm
