+++ pkg/providers/instance/errors.go
@@ -1,0 +1,43 @@
+/*
+Portions Copyright (c) Microsoft Corporation.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package instance
+
+import (
+	"fmt"
+)
+
+type InvalidInstanceStateError struct {
+	err          error
+	InstanceName string
+	Reason       string
+}
+
+func (e *InvalidInstanceStateError) Error() string {
+	return fmt.Sprintf("instance %s is in an invalid state for reason: %s", e.InstanceName, e.Reason)
+}
+
+func NewInvalidInstanceStateError(instanceName, failureReason string) error {
+	return &InvalidInstanceStateError{
+		InstanceName: instanceName,
+		Reason:       failureReason,
+	}
+}
+
+func IsInvalidInstanceStateError(err error) bool {
+	_, ok := err.(*InvalidInstanceStateError)
+	return ok
+}
