+++ pkg/providers/instance/instance.go
@@ -242,33 +242,6 @@ func (p *DefaultProvider) DeleteNic(ctx context.Context, nicName string) error {
 	return deleteNicIfExists(ctx, p.azClient.networkInterfacesClient, p.resourceGroup, nicName)
 }
 
-// createAKSIdentifyingExtension attaches a VM extension to identify that this VM participates in an AKS cluster
-func (p *DefaultProvider) createAKSIdentifyingExtension(ctx context.Context, vmName string) (err error) {
-	vmExt := p.getAKSIdentifyingExtension()
-	vmExtName := *vmExt.Name
-	log.FromContext(ctx).V(1).Info(fmt.Sprintf("Creating virtual machine AKS identifying extension for %s", vmName))
-	v, err := createVirtualMachineExtension(ctx, p.azClient.virtualMachinesExtensionClient, p.resourceGroup, vmName, vmExtName, *vmExt)
-	if err != nil {
-		log.FromContext(ctx).Error(err, fmt.Sprintf("Creating VM AKS identifying extension for VM %q failed", vmName))
-		return fmt.Errorf("creating VM AKS identifying extension for VM %q, %w failed", vmName, err)
-	}
-	log.FromContext(ctx).V(1).Info(fmt.Sprintf("Created  virtual machine AKS identifying extension for %s, with an id of %s", vmName, *v.ID))
-	return nil
-}
-
-func (p *DefaultProvider) createCSExtension(ctx context.Context, vmName string, cse string, isWindows bool) (err error) {
-	vmExt := p.getCSExtension(cse, isWindows)
-	vmExtName := *vmExt.Name
-	log.FromContext(ctx).V(1).Info(fmt.Sprintf("Creating virtual machine CSE for %s", vmName))
-	v, err := createVirtualMachineExtension(ctx, p.azClient.virtualMachinesExtensionClient, p.resourceGroup, vmName, vmExtName, *vmExt)
-	if err != nil {
-		log.FromContext(ctx).Error(err, fmt.Sprintf("Creating VM CSE for VM %q failed", vmName))
-		return fmt.Errorf("creating VM CSE for VM %q, %w failed", vmName, err)
-	}
-	log.FromContext(ctx).V(1).Info(fmt.Sprintf("Created virtual machine CSE for %s, with an id of %s", vmName, *v.ID))
-	return nil
-}
-
 func (p *DefaultProvider) newNetworkInterfaceForVM(opts *createNICOptions) armnetwork.Interface {
 	var ipv4BackendPools []*armnetwork.BackendAddressPool
 	for _, poolID := range opts.BackendPools.IPv4PoolIDs {
@@ -488,8 +461,13 @@ func (p *DefaultProvider) createVirtualMachine(ctx context.Context, opts *create
 	// If any of these properties are modified, the existing vm will return a 409 status code "PropertyChangeNotAllowed".
 	// this results in create being blocked on the nodeclaim until liveness TTL is hit.
 	resp, err := p.azClient.virtualMachinesClient.Get(ctx, p.resourceGroup, opts.VMName, nil)
-	// If status == ok, we want to return the existing vmm
+	// If the VM Exists but without CSE then we want to delete the nodeclaim
+	// If status == ok and we have the required extensions, we want to return the existing vm
 	if err == nil {
+		if !p.requiredVMExtensionsInstalled(resp.VirtualMachine) {
+			return nil, NewInvalidInstanceStateError(opts.VMName, "vm extensions required by aks are missing")
+		}
+
 		return &resp.VirtualMachine, nil
 	}
 	// if status != ok, and for a reason other than we did not find the vm
@@ -497,6 +475,7 @@ func (p *DefaultProvider) createVirtualMachine(ctx context.Context, opts *create
 	if azErr != nil && (azErr.ErrorCode != "NotFound" && azErr.ErrorCode != "ResourceNotFound") {
 		return nil, fmt.Errorf("getting VM %q: %w", opts.VMName, err)
 	}
+
 	vm := newVMObject(opts)
 	log.FromContext(ctx).V(1).Info(fmt.Sprintf("Creating virtual machine %s (%s)", opts.VMName, opts.InstanceType.Name))
 
@@ -787,60 +766,6 @@ func GetCapacityType(instance *armcompute.VirtualMachine) string {
 		return PriorityToCapacityType[string(*instance.Properties.Priority)]
 	}
 	return ""
-}
-
-func (p *DefaultProvider) getAKSIdentifyingExtension() *armcompute.VirtualMachineExtension {
-	const (
-		vmExtensionType                  = "Microsoft.Compute/virtualMachines/extensions"
-		aksIdentifyingExtensionName      = "computeAksLinuxBilling"
-		aksIdentifyingExtensionPublisher = "Microsoft.AKS"
-		aksIdentifyingExtensionTypeLinux = "Compute.AKS.Linux.Billing"
-	)
-
-	vmExtension := &armcompute.VirtualMachineExtension{
-		Location: lo.ToPtr(p.location),
-		Name:     lo.ToPtr(aksIdentifyingExtensionName),
-		Properties: &armcompute.VirtualMachineExtensionProperties{
-			Publisher:               lo.ToPtr(aksIdentifyingExtensionPublisher),
-			TypeHandlerVersion:      lo.ToPtr("1.0"),
-			AutoUpgradeMinorVersion: lo.ToPtr(true),
-			Settings:                &map[string]interface{}{},
-			Type:                    lo.ToPtr(aksIdentifyingExtensionTypeLinux),
-		},
-		Type: lo.ToPtr(vmExtensionType),
-	}
-
-	return vmExtension
-}
-
-func (p *DefaultProvider) getCSExtension(cse string, isWindows bool) *armcompute.VirtualMachineExtension {
-	const (
-		vmExtensionType     = "Microsoft.Compute/virtualMachines/extensions"
-		cseNameWindows      = "windows-cse-agent-karpenter"
-		cseTypeWindows      = "CustomScriptExtension"
-		csePublisherWindows = "Microsoft.Compute"
-		cseVersionWindows   = "1.10"
-		cseNameLinux        = "cse-agent-karpenter"
-		cseTypeLinux        = "CustomScript"
-		csePublisherLinux   = "Microsoft.Azure.Extensions"
-		cseVersionLinux     = "2.0"
-	)
-
-	return &armcompute.VirtualMachineExtension{
-		Location: lo.ToPtr(p.location),
-		Name:     lo.ToPtr(lo.Ternary(isWindows, cseNameWindows, cseNameLinux)),
-		Type:     lo.ToPtr(vmExtensionType),
-		Properties: &armcompute.VirtualMachineExtensionProperties{
-			AutoUpgradeMinorVersion: lo.ToPtr(true),
-			Type:                    lo.ToPtr(lo.Ternary(isWindows, cseTypeWindows, cseTypeLinux)),
-			Publisher:               lo.ToPtr(lo.Ternary(isWindows, csePublisherWindows, csePublisherLinux)),
-			TypeHandlerVersion:      lo.ToPtr(lo.Ternary(isWindows, cseVersionWindows, cseVersionLinux)),
-			Settings:                &map[string]interface{}{},
-			ProtectedSettings: &map[string]interface{}{
-				"commandToExecute": cse,
-			},
-		},
-	}
 }
 
 func ConvertToVirtualMachineIdentity(nodeIdentities []string) *armcompute.VirtualMachineIdentity {
